+++ ChangeLog.markdown
@@ -7,6 +7,7 @@
 * Fixed select2 (v.4.x) with server side processing https://github.com/vedmack/yadcf/issues/216
 * Fixed select2 reset https://github.com/vedmack/yadcf/issues/232
 * Destroy third party plugins on datatable destroy https://github.com/vedmack/yadcf/issues/213
+* Prevent column sort when hitting Enter in filter https://github.com/vedmack/yadcf/issues/233
 * Some of additional closed issues https://github.com/vedmack/yadcf/issues/227 / https://github.com/vedmack/yadcf/issues/236